Move healthchecks and installs to `pre` action
