+++ action.yaml
@@ -80,5 +80,6 @@ inputs:
 
 runs:
   using: "node16"
+  pre: "dist/pre.js"
   main: "dist/main.js"
   post: "dist/post.js"
