+++ src/main.ts
@@ -1,22 +1,16 @@
-import process from 'process'
 import fs from 'fs'
+import process from 'process'
 import * as core from '@actions/core'
-import fetch from 'node-fetch'
-import * as github from './github'
-import {HealthCheck} from './healthcheck'
-import * as workspace from './workspace'
 import * as coursier from './coursier'
-import {type Logger} from './logger'
-import {Input} from './input'
-import {type HttpClient} from './http'
-import * as mill from './mill'
 import {type Files} from './files'
+import * as github from './github'
+import {Input} from './input'
+import {type Logger} from './logger'
 import {nonEmpty, NonEmptyString} from './types'
+import * as workspace from './workspace'
 
 /**
  * Runs the action main code. In order it will do the following:
- * - Check connection with Maven Central
- * - Install Coursier
  * - Recover user inputs
  * - Get authenticated user data from provided Github Token
  * - Prepare Scala Steward's workspace
@@ -25,17 +19,8 @@ import {nonEmpty, NonEmptyString} from './types'
 async function run(): Promise<void> {
   try {
     const logger: Logger = core
-    const httpClient: HttpClient = {run: async url => fetch(url)}
     const files: Files = fs
     const inputs = Input.from(core, files, logger).all()
-    const healthCheck: HealthCheck = HealthCheck.from(logger, httpClient)
-
-    await healthCheck.mavenCentral()
-
-    await coursier.selfInstall()
-    await coursier.install('scalafmt')
-    await coursier.install('scalafix')
-    await mill.install()
 
     const user = await github.getAuthUser(inputs.github.token)
